usePlacesAutocomplete: Moved file location
Move the usePlacesAutocomplete JS and CSS files to the components folder. They comprise a component rather than a utility function. Import path corrected.
